Updated user/assetProductDtos and fixed checkout hooks

+++ server/StockLinx.Service/Services/AccessoryService.cs
@@ -223,42 +223,71 @@ namespace StockLinx.Service.Services
             return await _userProductRepository.GetDtoAsync(userProduct);
         }
 
-        public async Task<UserProductDto> CheckOutAsync(UserProductCheckOutDto checkOutDto)
+        public async Task<List<UserProductDto>> CheckOutAsync(UserProductCheckOutDto checkOutDto)
         {
+            List<UserProduct> userProducts = new List<UserProduct>();
             UserProduct userProduct = await _userProductRepository.GetByIdAsync(
                 checkOutDto.UserProductId
             );
-            Accessory accessory = await GetByIdAsync(checkOutDto.ProductId);
+            Accessory accessory = await GetByIdAsync((Guid)userProduct.AccessoryId);
+            bool isUserChanged = checkOutDto.UserId != null && checkOutDto.UserId != userProduct.UserId;
             switch (userProduct.Quantity - checkOutDto.Quantity)
             {
                 case 0:
-                    _userProductRepository.Remove(userProduct);
-                    await _customLogService.CreateCustomLog(
-                        "CheckOut",
-                        "Accessory",
-                        accessory.Id,
-                        accessory.Name,
-                        checkOutDto.Notes ?? "Checked out " + checkOutDto.Quantity + " units"
-                    );
+                    await CreateCheckLogAsync("CheckOut", accessory, checkOutDto.Quantity);
+                    if (isUserChanged)
+                    {
+                        userProduct.UserId = (Guid)checkOutDto.UserId;
+                        _userProductRepository.Update(userProduct, userProduct);
+                        await CreateCheckLogAsync(
+                            "CheckOut",
+                            accessory,
+                            await _userService.GetByIdAsync((Guid)checkOutDto.UserId),
+                            checkOutDto.Quantity
+                        );
+                        userProducts.Add(userProduct);
+                    }
+                    else
+                    {
+                        await CreateCheckLogAsync("CheckOut", accessory, checkOutDto.Quantity);
+                        _userProductRepository.Remove(userProduct);
+                    }
                     await _unitOfWork.CommitAsync();
-                    return null;
+                    ;
+                    return await _userProductRepository.GetDtosAsync(userProducts);
                 case > 0:
                     throw new Exception(
                         "Quantity must be less than or equal to the quantity in stock"
                     );
                 case < 0:
-                    UserProduct newUserProduct = userProduct;
-                    newUserProduct.Quantity -= checkOutDto.Quantity;
-                    _userProductRepository.Update(userProduct, newUserProduct);
-                    await _customLogService.CreateCustomLog(
-                        "CheckOut",
-                        "Accessory",
-                        accessory.Id,
-                        accessory.Name,
-                        checkOutDto.Notes ?? "Checked out " + checkOutDto.Quantity + " units"
-                    );
+                    UserProduct updatedUserProduct = userProduct;
+                    updatedUserProduct.Quantity -= checkOutDto.Quantity;
+                    _userProductRepository.Update(userProduct, updatedUserProduct);
+                    await CreateCheckLogAsync("CheckOut", accessory, checkOutDto.Quantity);
+                    userProducts.Add(updatedUserProduct);
+                    if (isUserChanged)
+                    {
+                        UserProduct newUserProduct = new UserProduct
+                        {
+                            Id = Guid.NewGuid(),
+                            AccessoryId = accessory.Id,
+                            UserId = (Guid)checkOutDto.UserId,
+                            AssignDate = DateTime.UtcNow,
+                            CreatedDate = DateTime.UtcNow,
+                            Quantity = checkOutDto.Quantity,
+                            Notes = checkOutDto.Notes,
+                        };
+                        await CreateCheckLogAsync(
+                            "CheckOut",
+                            accessory,
+                            await _userService.GetByIdAsync((Guid)checkOutDto.UserId),
+                            checkOutDto.Quantity
+                        );
+                        await _userProductRepository.AddAsync(newUserProduct);
+                        userProducts.Add(newUserProduct);
+                    }
                     await _unitOfWork.CommitAsync();
-                    return await _userProductRepository.GetDtoAsync(newUserProduct);
+                    return await _userProductRepository.GetDtosAsync(userProducts);
             }
         }
 
@@ -287,6 +316,36 @@ namespace StockLinx.Service.Services
         {
             var result = await _filterService.FilterAsync(filter);
             return await _accessoryRepository.GetDtosAsync(result.ToList());
+        }
+
+        public async Task CreateCheckLogAsync(
+            string action,
+            Accessory accessory,
+            User user,
+            int quantity
+        )
+        {
+            await _customLogService.CreateCustomLog(
+                action,
+                "Accessory",
+                accessory.Id,
+                accessory.Name,
+                "User",
+                user.Id,
+                user.FirstName + user.LastName,
+                "Checked " + quantity + " units"
+            );
+        }
+
+        public async Task CreateCheckLogAsync(string action, Accessory accessory, int quantity)
+        {
+            await _customLogService.CreateCustomLog(
+                action,
+                "Accessory",
+                accessory.Id,
+                accessory.Name,
+                "Checked " + quantity + " units"
+            );
         }
     }
 }
